test(cast): align readme/conformance tests with truncating cast semantics
Two related test fixes following recent cast-behavior changes:

* readme_quick_example::readme_typeerror_variant_triggers — update the
  TypeError expectations to match the new cast rules:
  - Float64 -> Int64 now TRUNCATES a finite non-integer toward zero
    (3.5 -> Int64(3)) to match pandas (br-frankenpandas-qcutc); only a
    non-finite float (±inf) still raises LossyFloatToInt. Test now drives
    f64::INFINITY for the error case and asserts 3.5 -> Ok(Int64(3)).
  - astype(bool) now follows numpy/pandas truthiness (br-frankenpandas-cyi4h):
    bool(2) and bool(0.5) are both True rather than raising
    InvalidBoolInt/InvalidBoolFloat. Those variants are retained for legacy
    strict callers but no longer fire here, so the test asserts Ok(Bool(true)).

* live_oracle_dataframe_series_core — pass str_split_n as i64 to
  split_get/rsplit_get/split_regex_get, matching the signed split-count
  signature (negative n = "all remaining", pandas semantics).

Test-only changes that track already-landed library behavior. beads synced.

+++ crates/fp-conformance/src/tests/live_oracle_dataframe_series_core.rs
@@ -20111,7 +20111,7 @@ fn live_oracle_series_str_split_get_index_1() {
         .str()
         .split_get(
             fixture.str_split_pat.as_deref().expect("str_split_pat"),
-            fixture.str_split_n.expect("str_split_n"),
+            fixture.str_split_n.expect("str_split_n") as i64,
         )
         .expect("str split_get");
     super::compare_series_expected(&actual, &expected).expect("pandas parity");
@@ -20215,7 +20215,7 @@ fn live_oracle_series_str_rsplit_get_index_0() {
         .str()
         .rsplit_get(
             fixture.str_split_pat.as_deref().expect("str_split_pat"),
-            fixture.str_split_n.expect("str_split_n"),
+            fixture.str_split_n.expect("str_split_n") as i64,
         )
         .expect("str rsplit_get");
     super::compare_series_expected(&actual, &expected).expect("pandas parity");
@@ -20318,7 +20318,7 @@ fn live_oracle_series_str_split_regex_get_first() {
         .str()
         .split_regex_get(
             fixture.regex_pattern.as_deref().expect("regex_pattern"),
-            fixture.str_split_n.expect("str_split_n"),
+            fixture.str_split_n.expect("str_split_n") as i64,
         )
         .expect("str split_regex_get");
     super::compare_series_expected(&actual, &expected).expect("pandas parity");
